fix: solve duplaction tags problem
