+++ packages/crd-scripts/src/conf/getPrerenderRoutes.js
@@ -1,26 +1,3 @@
-const { directoryTree } = require('./node-directory-tree')
-
-const getDirTree = (cmd) => {
-  const dir = cmd.markdownPaths
-  const dirs = Array.isArray(dir) ? dir : [dir]
-  const otherProps = {
-    mdconf: true,
-    extensions: /\.md/,
-    prerender: true,
-  }
-  const mapTagsWithArticle = []
-  const dirTree = dirs.map(path => directoryTree({
-    path,
-    options: otherProps,
-    mapTagsWithArticle
-  }))
-  return {
-    dirTree,
-    // map tags with path. [{ tagName: 'custom Tag 1', mapArticle: [{ path, name }]}]
-    mapTagsWithArticle
-  }
-}
-
 // eg: ['docs/quick_start.md', 'a']
 // output: ['/quick_start', '/a/b', '/a/b/c']
 const getPrerenderRoutes = (dirTree) => {
@@ -70,6 +47,5 @@ function recursive(
 }
 
 module.exports = {
-  getDirTree,
   getPrerenderRoutes
 }
